src/cv/data/refactor-price-tag-detector/pyproject.toml
[project]
name = "price-tag-detector"
version = "1.0.0"
description = "價格標籤檢測系統 - 基於 YOLO + 顏色過濾 + OCR"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # YOLO 物體檢測
    "ultralytics>=8.4.0",
    # 深度學習框架
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    # 圖像處理
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    # OCR 文字識別
    "pytesseract>=0.3.10",
    # 進度條
    "tqdm>=4.66.0",
    # 配置管理
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    # 測試
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",

    # 程式碼品質
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "cli", "utils"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=src",
    "--cov-report=term-missing",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
